+++ lib/src/quran/presentation/widgets/auto_scroll/auto_scroll_speed_slider.dart
@@ -8,15 +8,19 @@ class AutoScrollSpeedSlider extends StatelessWidget {
   const AutoScrollSpeedSlider({
     super.key,
     required this.isDark,
+    required this.autoScrollStyle,
+    required this.languageCode,
   });
 
   final bool isDark;
+  final AutoScrollStyle autoScrollStyle;
+  final String languageCode;
 
   @override
   Widget build(BuildContext context) {
     final autoScrollCtrl = AutoScrollCtrl.instance;
-    final style = AutoScrollTheme.of(context)?.style ??
-        AutoScrollStyle.defaults(isDark: isDark, context: context);
+    final style = AutoScrollTheme.of(context)?.style ?? autoScrollStyle;
+    AutoScrollStyle.defaults(isDark: isDark, context: context);
 
     return Obx(() {
       if (!autoScrollCtrl.state.isActive.value) {
@@ -93,7 +97,8 @@ class AutoScrollSpeedSlider extends StatelessWidget {
                 SizedBox(
                   width: 36,
                   child: Text(
-                    speed.toStringAsFixed(1),
+                    speed.toStringAsFixed(1).convertNumbersAccordingToLang(
+                        languageCode: languageCode),
                     textAlign: TextAlign.center,
                     style: style.speedLabelStyle ??
                         TextStyle(